Initial Upload
Althought he branch was originally made, there was an issue with connecting my computer's folders to GitHub. This was found to be due to a syncing issue with OneDrive, thus a lot of the work is uploaded in a single initial chunk for exercise 2. This will be rectified for Exercises 3-4.